requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
python-dotenv

langchain-core
langchain-community
langchain-groq
groq

langgraph
langgraph-checkpoint-sqlite

httpx
aiofiles
sqlite-utils
